+++ src/main/scala/com/lucendar/strm/common/strm/OpenStrmResult.java
@@ -5,6 +5,7 @@ import java.util.StringJoiner;
 public class OpenStrmResult {
 
     private String reqId;
+    private int channel;
     private boolean ctrl;
     private boolean ready;
     private String mediaTyp;
@@ -24,12 +25,13 @@ public class OpenStrmResult {
     public OpenStrmResult() {
     }
 
-    public OpenStrmResult(String reqId, boolean ctrl, boolean ready, String mediaTyp, String playUrl, String wsUrl,
+    public OpenStrmResult(String reqId, int channel, boolean ctrl, boolean ready, String mediaTyp, String playUrl, String wsUrl,
                           String taUrl,
                           int keepIntv, String host, int port, boolean tcp,
                           int port2019, boolean tcp2019, Boolean recording, AudioConfig audioConfig,
                           Boolean detectMediaTyp) {
         this.reqId = reqId;
+        this.channel = channel;
         this.ctrl = ctrl;
         this.ready = ready;
         this.mediaTyp = mediaTyp;
@@ -53,6 +55,14 @@ public class OpenStrmResult {
 
     public void setReqId(String reqId) {
         this.reqId = reqId;
+    }
+
+    public int getChannel() {
+        return channel;
+    }
+
+    public void setChannel(int channel) {
+        this.channel = channel;
     }
 
     public boolean isCtrl() {
@@ -177,23 +187,24 @@ public class OpenStrmResult {
 
     @Override
     public String toString() {
-        return new StringJoiner(", ", OpenStrmResult.class.getSimpleName() + "[", "]")
-                .add("reqId='" + reqId + "'")
-                .add("ctrl=" + ctrl)
-                .add("ready=" + ready)
-                .add("mediaTyp='" + mediaTyp + "'")
-                .add("playUrl='" + playUrl + "'")
-                .add("wsUrl='" + wsUrl + "'")
-                .add("taUrl='" + taUrl + "'")
-                .add("keepIntv=" + keepIntv)
-                .add("host='" + host + "'")
-                .add("port=" + port)
-                .add("tcp=" + tcp)
-                .add("port2019=" + port2019)
-                .add("tcp2019=" + tcp2019)
-                .add("recording=" + recording)
-                .add("audioConfig=" + audioConfig)
-                .add("detectMediaTyp=" + detectMediaTyp)
-                .toString();
+        return "OpenStrmResult{" +
+                "reqId='" + reqId + '\'' +
+                ", channel=" + channel +
+                ", ctrl=" + ctrl +
+                ", ready=" + ready +
+                ", mediaTyp='" + mediaTyp + '\'' +
+                ", playUrl='" + playUrl + '\'' +
+                ", wsUrl='" + wsUrl + '\'' +
+                ", taUrl='" + taUrl + '\'' +
+                ", keepIntv=" + keepIntv +
+                ", host='" + host + '\'' +
+                ", port=" + port +
+                ", tcp=" + tcp +
+                ", port2019=" + port2019 +
+                ", tcp2019=" + tcp2019 +
+                ", recording=" + recording +
+                ", audioConfig=" + audioConfig +
+                ", detectMediaTyp=" + detectMediaTyp +
+                '}';
     }
 }